A few random updates to make things less horrifyingly out of date.
Delete some advice that can never, ever have worked.

There are a couple of XXX comments for bits I don't know how to
update.  It would be really good not to release Python 2.5 with these
in place :)

This file is way too big.  There's basically no chance of it staying
up to date.